+++ sdk/spring/spring-cloud-azure-resourcemanager/src/test/java/com/azure/spring/cloud/resourcemanager/implementation/provisioning/DefaultServiceBusProvisionerTests.java
@@ -18,6 +18,8 @@ import com.azure.spring.cloud.core.properties.resource.AzureResourceMetadata;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.ServiceBusQueueCrud;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.ServiceBusTopicCrud;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.ServiceBusTopicSubscriptionCrud;
+import com.azure.spring.cloud.stream.binder.servicebus.core.properties.ServiceBusConsumerProperties;
+import com.azure.spring.cloud.stream.binder.servicebus.core.properties.ServiceBusProducerProperties;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import reactor.util.function.Tuples;
@@ -149,17 +151,17 @@ class DefaultServiceBusProvisionerTests {
 
         @Override
         public void provisionQueue(String namespace, String queue) {
-            this.queueCrud.getOrCreate(Tuples.of(namespace, queue));
+            this.queueCrud.getOrCreate(Tuples.of(namespace, queue, new ServiceBusProducerProperties()));
         }
 
         @Override
         public void provisionTopic(String namespace, String topic) {
-            this.topicCrud.getOrCreate(Tuples.of(namespace, topic));
+            this.topicCrud.getOrCreate(Tuples.of(namespace, topic, new ServiceBusProducerProperties()));
         }
 
         @Override
         public void provisionSubscription(String namespace, String topic, String subscription) {
-            this.subscriptionCrud.getOrCreate(Tuples.of(namespace, topic, subscription));
+            this.subscriptionCrud.getOrCreate(Tuples.of(namespace, topic, subscription, new ServiceBusConsumerProperties()));
         }
     }
 }